requirements.txt
Pillow
cloudinary
aiohttp
cairosvg
vharfbuzz
gidgethub
pygit2
